labels for players name
I used eclipse instead of netbeans because I cannot adjust the size of the image in netbeans. However, I'm still using netbeans pallet to organize the pallets. I added the labels for the players' name, if you don't like any of the name you can change it. For the move button, there's an action listener for it. To get the x and y you can just use name1.getX() or name1.getY()